Added wgetall
This script will download all files in a particular html index.  This
can be very useful but the exact formatting will depend on the server
you are downloading from.  This just happens to work best for my uses.
Source:http://stackoverflow.com/questions/273743/using-wget-to-recursive
ly-fetch-a-directory-with-arbitrary-files-in-it
Note: this does rely on having wget installed or available in your $PATH